% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllWhalesAvgs}
\alias{AllWhalesAvgs}
\title{Lunge Kinematics and Energetics Whale Average Values}
\format{
A data frame with 42 rows and 13 (for now) variables:
\describe{
\item{Species}{The species of that particular whale}
\item{whaleName}{That particular whale's ID}
}
}
\source{
\url{http://www.FIXME.info/}
}
\usage{
AllWhalesAvgs
}
\description{
Values from AllWhalesLungeTableTrunc have been consolidated into mean values.
}
\keyword{datasets}
