% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllWhalesAvgsWithFilt}
\alias{AllWhalesAvgsWithFilt}
\title{Lunge Kinematics and Energetics Whale Average Values + Filtration Data}
\format{
A data frame with 28 rows and 10 (for now) variables:
\describe{
\item{Species}{The species of that particular whale}
\item{whaleName}{That particular whale's ID}
}
}
\source{
\url{http://www.FIXME.info/}
}
\usage{
AllWhalesAvgsWithFilt
}
\description{
Values from AllWhalesLungeTableTrunc have been consolidated into mean values and filtration data has been added from Shirel. Lunges without filtration data have been filtered out of the dateset.
}
\keyword{datasets}
