% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllWhalesLungeTableTrunc}
\alias{AllWhalesLungeTableTrunc}
\title{Lunge Kinematics and Energetics Dataset For Gough et al. 2021}
\format{
A data frame with 1260 rows and 56 (for now) variables:
\describe{
\item{LungeI}{Data index of a lunge in the deployment's PRH file}
\item{timeoflunge}{time of a lunge, local}
}
}
\source{
\url{http://www.FIXME.info/}
}
\usage{
AllWhalesLungeTableTrunc
}
\description{
A dataset containing kinematics and stuff...
}
\keyword{datasets}
